pyproject.toml
[tool.poetry]
name = "appium-hugo"
version = "0.1.0"
description = ""
authors = ["hei.cga"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.9"
appium-python-client = "^4.1.0"
ignr = "^2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
